requirements.txt
discord.py[voice]
PyNaCl
opuslib
aiohttp
python-dotenv